+++ servicelog/mapka2/output.go
@@ -37,7 +37,7 @@ type OutputRecord struct {
 	IsAnonymous bool                     `json:"isAnonymous"`
 	IsQuery     bool                     `json:"isQuery"`
 	GeoIP       servicelog.GeoDataRecord `json:"geoip,omitempty"`
-	ProcTime    float32                  `json:"procTime"`
+	ProcTime    float64                  `json:"procTime"`
 }
 
 // SetLocation sets all the location related properties
